@@ -1,9 +1,9 @@
 {
   "name": "react-redux-boilerplate",
-  "version": "0.7.14",
+  "version": "0.8.14",
   "description": "React Redux Boilerplate is a workflow boilerplate that make life easier for developers by providing a virtual development environment and production ready build process framework out of the box.",
   "scripts": {
-    "dev": "NODE_ENV=development webpack-dashboard -p 3300 -c red -- node bin/commands.js dev",
+    "dev": "NODE_ENV=development webpack-dashboard -p 3300 -c red -t dashboard -- node bin/commands.js dev",
     "build": "NODE_ENV=production node bin/commands.js build",
     "build:release": "NODE_ENV=release node bin/commands.js build",
     "test": "jest --verbose",
@@ -18,68 +18,71 @@
   "license": "MIT",
   "devDependencies": {
     "assets-webpack-plugin": "^3.5.1",
-    "babel-cli": "^6.24.0",
-    "babel-core": "^6.24.0",
-    "babel-eslint": "^7.2.1",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-eslint": "^7.2.3",
     "babel-jest": "^19.0.0",
-    "babel-loader": "^6.4.1",
+    "babel-loader": "^7.0.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-plugin-typecheck": "^3.9.0",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-latest": "^6.24.0",
-    "babel-preset-react": "^6.23.0",
+    "babel-preset-latest": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
-    "babel-preset-stage-0": "^6.22.0",
-    "babel-register": "^6.24.0",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-register": "^6.24.1",
     "babel-runtime": "^6.23.0",
     "browser-sync": "^2.18.8",
     "browser-sync-webpack-plugin": "^1.1.4",
-    "case-sensitive-paths-webpack-plugin": "^1.1.4",
+    "case-sensitive-paths-webpack-plugin": "^2.0.0",
     "colors": "^1.1.2",
     "concurrently": "^3.4.0",
     "config": "^1.25.1",
-    "css-loader": "^0.27.3",
-    "debug": "^2.6.3",
+    "css-loader": "^0.28.0",
+    "debug": "^2.6.6",
     "dotenv": "^4.0.0",
-    "eslint": "^3.18.0",
+    "eslint": "^3.19.0",
     "eslint-config-airbnb": "^14.1.0",
     "eslint-friendly-formatter": "^2.0.7",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
     "eslint-plugin-react": "^6.10.3",
     "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.10.1",
+    "file-loader": "^0.11.1",
     "html-webpack-plugin": "^2.28.0",
     "image-webpack-loader": "^3.3.0",
     "jest-cli": "^19.0.2",
-    "jsonfile": "^2.4.0",
+    "jsonfile": "^3.0.0",
     "postcss-cssnext": "^2.10.0",
     "postcss-import": "^9.1.0",
     "postcss-loader": "^1.3.3",
-    "postcss-nested": "^1.0.0",
+    "postcss-nested": "^1.0.1",
     "precss": "^1.4.0",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-router": "^3.0.2",
+    "react-router": "^4.1.1",
     "shelljs": "^0.7.7",
-    "style-loader": "^0.16.0",
-    "webpack": "^2.3.2",
+    "style-loader": "^0.16.1",
+    "webpack": "^2.4.1",
     "webpack-bundle-analyzer": "^2.4.0",
-    "webpack-dashboard": "^0.3.0",
-    "webpack-dev-server": "^2.4.2",
+    "webpack-dashboard": "^0.4.0",
+    "webpack-dev-server": "^2.4.5",
     "webpack-s3-plugin": "^1.0.0-rc.0"
   },
   "dependencies": {
+    "history": "^4.6.1",
     "immutable": "^3.8.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-helmet": "^5.0.1",
-    "react-redux": "^5.0.3",
-    "react-router": "^3.0.2",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
+    "react-helmet": "^5.0.3",
+    "react-redux": "^5.0.4",
+    "react-responsive": "^1.2.10",
+    "react-router": "^4.1.1",
+    "react-router-dom": "^4.1.1",
     "react-router-redux": "^4.0.8",
     "redux": "^3.6.0",
-    "redux-actions": "^2.0.1",
-    "redux-form": "^6.6.0",
+    "redux-actions": "^2.0.2",
+    "redux-form": "^6.6.3",
     "redux-promised": "^1.3.0",
     "reselect": "^3.0.0"
   },
